+++ internal/pkg/fetch/fetch_fuzz_test.go
@@ -1,6 +1,6 @@
 // -----------------------------------------------------------------------------
 // SPDX-FileCopyrightText: Copyright © 2024 bomctl a Series of LF Projects, LLC
-// SPDX-FileName: internal/pkg/push/push_fuzz.go
+// SPDX-FileName: internal/pkg/fetch/fetch_fuzz.go
 // SPDX-FileType: SOURCE
 // SPDX-License-Identifier: Apache-2.0
 // -----------------------------------------------------------------------------
@@ -17,30 +17,26 @@
 // limitations under the License.
 // -----------------------------------------------------------------------------
 
-package push
+package fetch_test
 
 import (
 	"context"
 	"testing"
 
+	"github.com/bomctl/bomctl/internal/pkg/fetch"
 	"github.com/bomctl/bomctl/internal/pkg/options"
 )
 
-const (
-	id      = "urn:uuid:f360ad8b-dc41-4256-afed-337a04dff5db"
-	pushURL = "test.com"
-)
+const testURL = "https://raw.githubusercontent.com/bomctl/bomctl-playground/main/bomctl_bomctl_v0.3.0.cdx.json"
 
-func FuzzPush(f *testing.F) {
-	f.Add(id, pushURL)
+func FuzzFetch(f *testing.F) {
+	f.Add([]byte(testURL))
 
-	f.Fuzz(func(t *testing.T, id, pushURL string) {
-		ctx := context.Background()
-		opts := options.Options{}
-		opts.WithContext(ctx)
-		pOpts := options.PushOptions{Options: &opts}
-
-		err := Push(id, pushURL, &pOpts)
+	f.Fuzz(func(t *testing.T, data []byte) {
+		_, err := fetch.Fetch(
+			string(data),
+			&options.FetchOptions{Options: options.New().WithContext(context.Background())},
+		)
 
 		if err == nil {
 			t.Errorf("%s", err)
