+++ internal/pkg/push/push_fuzz_test.go
@@ -1,6 +1,6 @@
 // -----------------------------------------------------------------------------
 // SPDX-FileCopyrightText: Copyright © 2024 bomctl a Series of LF Projects, LLC
-// SPDX-FileName: internal/pkg/fetch/fetch_fuzz.go
+// SPDX-FileName: internal/pkg/push/push_fuzz.go
 // SPDX-FileType: SOURCE
 // SPDX-License-Identifier: Apache-2.0
 // -----------------------------------------------------------------------------
@@ -17,27 +17,30 @@
 // limitations under the License.
 // -----------------------------------------------------------------------------
 
-package fetch
+package push_test
 
 import (
 	"context"
 	"testing"
 
 	"github.com/bomctl/bomctl/internal/pkg/options"
+	"github.com/bomctl/bomctl/internal/pkg/push"
 )
 
-const testURL = "https://raw.githubusercontent.com/bomctl/bomctl-playground/main/bomctl_bomctl_v0.3.0.cdx.json"
+const (
+	id      = "urn:uuid:f360ad8b-dc41-4256-afed-337a04dff5db"
+	pushURL = "test.com"
+)
 
-func FuzzFetch(f *testing.F) {
-	f.Add([]byte(testURL))
+func FuzzPush(f *testing.F) {
+	f.Add(id, pushURL)
 
-	f.Fuzz(func(t *testing.T, data []byte) {
-		ctx := context.Background()
-		opts := options.Options{}
-		opts.WithContext(ctx)
-		fOpts := options.FetchOptions{Options: &opts}
-
-		_, err := Fetch(string(data), &fOpts)
+	f.Fuzz(func(t *testing.T, id, pushURL string) {
+		err := push.Push(
+			id,
+			pushURL,
+			&options.PushOptions{Options: options.New().WithContext(context.Background())},
+		)
 
 		if err == nil {
 			t.Errorf("%s", err)
